Update(fix) wallet-server.md
Updated HW requirements, I don't think it was possible to get hub-server running at all on the old values. New values used contain some amount of guessing.

The current steps left room for a lot of guessing of what exactly to do.(Like when to exactly start lbrycrdd, or the info that it doesn't matter when different things are started/stopped) This wasn't completely obvious for me. I added bit more details to the guide, which hopefully makes it easier to get hub running for people.

As version 17.4.6 requires much less RAM, I added it to the guide. I don't know if there are other things than RAM, for people to care, when they choose the version to use. Probably could use some piece of text for how to choose?

Updated word "wallet server" to "hub server" when referring to whole thing.

Experience with old guide:
It took me ~8 days to get hub-server running, mainly because lbrycrd crashing for lack of RAM(had 16GB) and no info about blockchain snapshot on this guide(I did --reindex instead).   
After needing to ask help and trying with 17.4.6, it took me ~2 days to get everything running. And that was when letting lbrycrd fully sync before starting docker stuff( I didn't want more unexpected crashes that make lbrycrd want --reindex). In the end, on 17.4.6 it was very straight-forward to get things going.  (Luckily I had 315GB of storage on server, instead of 200GB as in guide)